fix: active_flag 타입을 Byte로 변경 (MySQL TINYINT 매핑)

+++ src/main/java/com/bookchigi/user/domain/User.java
@@ -60,7 +60,7 @@ public class User {
      * insert/update 불가
      */
     @Column(name = "active_flag", insertable = false, updatable = false)
-    private Integer activeFlag;
+    private Byte activeFlag;
 
     /* =========================
        생성 메서드
